pyproject.toml
[tool.poetry]
name = "snow-challenge"
version = "0.1.0"
description = ""
authors = ["guitabalipa <guilherme.tabalipa@snowmanlabs.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.0"
uvicorn = "^0.11.8"
sqlalchemy = "^1.3.19"
alembic = "^1.4.2"
psycopg2 = "^2.8.5"
python-multipart = "^0.0.5"
python-jose = "^3.2.0"
passlib = "^1.7.2"
bcrypt = "^3.2.0"
cryptography = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
